Restructured directory and rewrote CMakeLists
