+++ src/board_cells_delimitation/board_cells_delimitation.cpp
