+++ src/board_cells_delimitation/board_cells_delimitation.h
@@ -16,7 +16,7 @@
 #include <QFileDialog>
 #include <QApplication>
 
-#include "tictactoe_utils.h"
+#include "baxterTictactoe/tictactoe_utils.h"
 
 namespace ttt
 {
