+++ src/concat_trajectory_xml_files/concat_trajectory_xml_files.cpp
@@ -1,4 +1,4 @@
-#include "trajectory_xml_parser.h"
+#include "trajectory_xml_parser/trajectory_xml_parser.h"
 
 int main(int argc, char ** argv)
 {
