+++ src/move_maker_simpleclient/move_maker_simpleclient.cpp
