+++ src/test_trajectory_xml_parser/test_trajectory_xml_parser.cpp
@@ -1,4 +1,4 @@
-#include "trajectory_xml_parser.h"
+#include "trajectory_xml_parser/trajectory_xml_parser.h"
 
 int main(int argc, char ** argv)
 {
