+++ src/trajectory_from_raw_to_xml/trajectory_from_raw_to_xml.cpp
@@ -1,6 +1,6 @@
 #include <ros/console.h>
 
-#include "trajectory_xml_parser.h"
+#include "trajectory_xml_parser/trajectory_xml_parser.h"
 
 int main(int argc, char ** argv)
 {
